Move markdown description cells from other notebook here
Reason to move them is that this way this notebook can be solely for
experimentation to see how code works and describe everything. Then
there is a final copy python file that can be run to deliver final
results. This way all the messy experimentation is centralized in an
environment tolerant of lots of different snippets of code (ipython
notebook's purpose). And the final copy code revisions can be easily
tracked an modified as they are located in a python file.